Add GitHub Actions workflow for G4™ UiaPeek release
This commit introduces a new project named ".github" in the G4.sln solution file, which includes a GitHub Actions workflow defined in `release-pipline.yml`. The workflow automates versioning, building, publishing, and creating a GitHub release for the G4™ UiaPeek project. Key steps include checking out the repository, parsing and validating the build version, restoring packages, building the project, creating a build artifact, and uploading it, along with creating a GitHub release with the generated artifact.